enabled filtering of progress meters; renamed pages
