+++ apps/cranio-provider/src/views/cs-all-general.vue
