+++ apps/cranio-provider/src/views/cs-all-surgical.vue
