+++ apps/cranio-provider/src/views/cs-center-general.vue
